[build_actor] Add skeleton and animation support (#3638)
This adds a feature to `build_actor` to support importing skeletons and
animations from .glb files.

Multiple animations are handled and will use the name in the GLB. The
default `viewer` process will end up playing back the first animation.

There are a few limitations:
- You can only have around 100 bones. It is technically possibly to have
slightly more, but certain animations may fail to compress when there
are more than ~100 bones.
- Currently, all animations have 60 keyframes per second. This is a
higher quality than what is normally used. If animation size becomes
problematic, we could make this customizable somehow.
- There is no support for the `align` bone.

---------

+++ decompiler/level_extractor/merc_replacement.cpp
@@ -99,9 +99,9 @@ const tfrag3::MercVertex& find_closest(const std::vector<tfrag3::MercVertex>& ol
                                        float y,
                                        float z) {
   float best_dist = 1e10;
-  int best_idx = 0;
+  size_t best_idx = 0;
 
-  for (int i = 0; i < old.size(); i++) {
+  for (size_t i = 0; i < old.size(); i++) {
     auto& v = old[i];
     float dx = v.pos[0] - x;
     float dy = v.pos[1] - y;
